container-timing/tentative/nested-containertiming-and-ignore-on-img.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Container Timing: an image with both containertiming and containertiming-ignore inside an outer containertiming root</title>
<body>
<style>
body {
  margin: 0;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/container-timing/resources/container-timing-helpers.js"></script>
<script>
  let beforeRender;
  async_test(function (t) {
    assert_implements(window.PerformanceContainerTiming, "PerformanceContainerTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        // Only one entry: the inner image root. The outer containertiming
        // root must NOT receive the paint because the image carries
        // containertiming-ignore which blocks upward propagation.
        assert_equals(entryList.getEntries().length, 1,
            'one entry expected for the image root, none for the outer root');
        const entry = entryList.getEntries()[0];
        checkContainerEntry(entry, 'image_ct', 'my_id', beforeRender);
        checkContainerIntersectionRect(entry, [0, 100, 0, 100])
        checkContainerSize(entry, 10000);
      })
    );
    observer.observe({entryTypes: ['container']});

    // Outer containertiming root.
    const outer = document.createElement('div');
    outer.setAttribute('containertiming', 'outer_ct');
    document.body.appendChild(outer);

    // Image that is itself a container timing root and has
    // containertiming-ignore set. The ignore must prevent the image's paint
    // from propagating to the outer containertiming root, but the image
    // itself must still report under its own root.
    const img = document.createElement('img');
    img.src = '/container-timing/resources/square100.png';
    img.setAttribute('containertiming', 'image_ct');
    img.setAttribute('containertiming-ignore', '');
    img.setAttribute('id', 'my_id');
    outer.appendChild(img);

    beforeRender = performance.now();
  }, 'Image with containertiming + containertiming-ignore inside outer containertiming reports under itself only.');
</script>

</body>
